README
To make changes, simply open up the Repo in file explorer and go about
your business.  when you are finished, all you have to do is open the
GitHub GUI and write a report on what you changed (this is required).
This will keep code from getting overwritten and breaking other people's
functions.