Reverted back.
Housing was not complete and broke the mission file.  Please only commit
changes when you have tested them and know that they are working.

If you are going to commit unfinished work, please state in the
description that it is not finished so it does not get mistaken as
broken code.

All files have been backed up (see Jonsandman)

Thank You,

Skeith