fixed some errors
-found more uses of the FAR script that was added from the merge
-deleted revive and revive action because they were outdated and moved
-player now gets ejected if they are in a vehicle during unconsciousness
(can be changed if we want to PUT players in vehicles ie: ambulance)
-added a function in the abortenabled for deleting unconscious markers
if players try to abort, it is untested so I commented it out, will need
someone to test it with me next time I am on.

Please be careful when updating.  make sure you review what you commit
so that no mistakes are being made.