Fixes
-lock picking on heli’s is working, it should be turned off. (if you are
going to do it, make the chance lower) Disabled it again
-No Steel Proccessing License; Reworked the addaction
-Cant collect Lumber; Alan's Hardwood Supply was incorrectly marked,
fixed it up. Lumber requires a chainsaw to log.
-Medics spawn in with different money than usual. Y and U do no work.
Looked fine when I just checked it
-When spawning in as a medic it said it had already tried to connect me
3 times and I have to respawn to join. I had this issue on my local
server, it was the MySQL side rather than a scripting issue
- If you escape from jail you shouldn’t lose your license. Difficult to
handle properly as it removes it when jailed. I have removed losing
Rebel Training entirely from the jailing process for now. This can be
re-added later if required.
-2 DMV in Athira, one is unmarked. Removed the random knockoff license
guy in Athira
-Cops aren’t getting time menu when getting sent to jail. *SHOULD* be
fixed now, wasn't calling the dialog option