[SKEITH HELP ME PLEHZ] Storage
Boxes can be placed by the player, and correctly store information in
the "storage" section of the houses table Structure is : "ID,
BoxType,BoxModel,XYZ".

I am having serious trouble reading the XYZ co-ordinates onto the
initHouses.sqf so it spawns the boxes at the position set in the
Database. Any help would be appreciated